Move configuration implement in memory to separate package
