+++ core/src/main/java/xyz/tozymc/configuration/memory/MemoryStorage.java
@@ -1,4 +1,4 @@
-package xyz.tozymc.configuration;
+package xyz.tozymc.configuration.memory;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
